[CP-2608] phrase.js configure moved to scripts
