+++ scripts/sync-translations/phrase.js
